requirements.txt
PyMuPDF>=1.24.0
Pillow>=10.0.0
boto3>=1.34.0
